+++ packages/skift.hjert.kernel/includes/kernel/tasking_new.h
@@ -1,0 +1,6 @@
+#pragma once
+
+/* Copyright © 2018-2019 MAKER.                                               */
+/* This code is licensed under the MIT License.                               */
+/* See: LICENSE.md                                                            */
+